Rename files in src/parse/, mark class private members as such
