+++ src/compose/compose-collection.ts
@@ -1,10 +1,6 @@
 import { Node, Scalar, YAMLMap, YAMLSeq } from '../ast/index.js'
 import type { Document } from '../doc/Document.js'
-import type {
-  BlockMap,
-  BlockSequence,
-  FlowCollection
-} from '../parse/parser.js'
+import type { BlockMap, BlockSequence, FlowCollection } from '../parse/tokens.js'
 import { resolveBlockMap } from './resolve-block-map.js'
 import { resolveBlockSeq } from './resolve-block-seq.js'
 import { resolveFlowCollection } from './resolve-flow-collection.js'
