+++ src/compose/resolve-flow-collection.ts
@@ -1,7 +1,7 @@
 import { Node, Pair, YAMLMap, YAMLSeq } from '../ast/index.js'
 import { Type } from '../constants.js'
 import type { Document } from '../doc/Document.js'
-import type { FlowCollection, SourceToken, Token } from '../parse/parser.js'
+import type { FlowCollection, SourceToken, Token } from '../parse/tokens.js'
 import { composeNode } from './compose-node.js'
 import { resolveEnd } from './resolve-end.js'
 import { resolveMergePair } from './resolve-merge-pair.js'
