+++ src/parse/tokens.ts
@@ -1,0 +1,204 @@
+export type SourceTokenType =
+  | 'byte-order-mark'
+  | 'doc-mode'
+  | 'scalar'
+  | 'doc-start'
+  | 'doc-end'
+  | 'space'
+  | 'comment'
+  | 'newline'
+  | 'directive-line'
+  | 'alias'
+  | 'anchor'
+  | 'tag'
+  | 'seq-item-ind'
+  | 'explicit-key-ind'
+  | 'map-value-ind'
+  | 'flow-map-start'
+  | 'flow-map-end'
+  | 'flow-seq-start'
+  | 'flow-seq-end'
+  | 'flow-error-end'
+  | 'comma'
+  | 'single-quoted-scalar'
+  | 'double-quoted-scalar'
+  | 'block-scalar-header'
+
+export interface SourceToken {
+  type: Exclude<SourceTokenType, DocumentEnd['type'] | FlowScalar['type']>
+  indent: number
+  source: string
+}
+
+export interface ErrorToken {
+  type: 'error'
+  offset?: number
+  source: string
+  message: string
+}
+
+export interface Directive {
+  type: 'directive'
+  source: string
+}
+
+export interface Document {
+  type: 'document'
+  offset: number
+  start: SourceToken[]
+  value?: Token
+  end?: SourceToken[]
+}
+
+export interface DocumentEnd {
+  type: 'doc-end'
+  offset: number
+  source: string
+  end?: SourceToken[]
+}
+
+export interface FlowScalar {
+  type: 'alias' | 'scalar' | 'single-quoted-scalar' | 'double-quoted-scalar'
+  offset: number
+  indent: number
+  source: string
+  end?: SourceToken[]
+}
+
+export interface BlockScalar {
+  type: 'block-scalar'
+  offset: number
+  indent: number
+  props: Token[]
+  source?: string
+}
+
+export interface BlockMap {
+  type: 'block-map'
+  offset: number
+  indent: number
+  items: Array<
+    | { start: SourceToken[]; key?: never; sep?: never; value?: never }
+    | {
+        start: SourceToken[]
+        key: Token | null
+        sep: SourceToken[]
+        value?: Token
+      }
+  >
+}
+
+export interface BlockSequence {
+  type: 'block-seq'
+  offset: number
+  indent: number
+  items: Array<{ start: SourceToken[]; value?: Token; sep?: never }>
+}
+
+export interface FlowCollection {
+  type: 'flow-collection'
+  offset: number
+  indent: number
+  start: SourceToken
+  items: Array<Token>
+  end: SourceToken[]
+}
+
+export type Token =
+  | SourceToken
+  | ErrorToken
+  | Directive
+  | Document
+  | DocumentEnd
+  | FlowScalar
+  | BlockScalar
+  | BlockMap
+  | BlockSequence
+  | FlowCollection
+
+/** The byte order mark */
+export const BOM = '\u{FEFF}'
+
+/** Start of doc-mode */
+export const DOCUMENT = '\x02' // C0: Start of Text
+
+/** Unexpected end of flow-mode */
+export const FLOW_END = '\x18' // C0: Cancel
+
+/** Next token is a scalar value */
+export const SCALAR = '\x1f' // C0: Unit Separator
+
+export function prettyToken(token: string) {
+  switch (token) {
+    case BOM:
+      return '<BOM>'
+    case DOCUMENT:
+      return '<DOC>'
+    case FLOW_END:
+      return '<FLOW_END>'
+    case SCALAR:
+      return '<SCALAR>'
+    default:
+      return JSON.stringify(token)
+  }
+}
+
+export function tokenType(source: string): SourceTokenType | null {
+  switch (source) {
+    case BOM:
+      return 'byte-order-mark'
+    case DOCUMENT:
+      return 'doc-mode'
+    case FLOW_END:
+      return 'flow-error-end'
+    case SCALAR:
+      return 'scalar'
+    case '---':
+      return 'doc-start'
+    case '...':
+      return 'doc-end'
+    case '':
+    case '\n':
+    case '\r\n':
+      return 'newline'
+    case '-':
+      return 'seq-item-ind'
+    case '?':
+      return 'explicit-key-ind'
+    case ':':
+      return 'map-value-ind'
+    case '{':
+      return 'flow-map-start'
+    case '}':
+      return 'flow-map-end'
+    case '[':
+      return 'flow-seq-start'
+    case ']':
+      return 'flow-seq-end'
+    case ',':
+      return 'comma'
+  }
+  switch (source[0]) {
+    case ' ':
+    case '\t':
+      return 'space'
+    case '#':
+      return 'comment'
+    case '%':
+      return 'directive-line'
+    case '*':
+      return 'alias'
+    case '&':
+      return 'anchor'
+    case '!':
+      return 'tag'
+    case "'":
+      return 'single-quoted-scalar'
+    case '"':
+      return 'double-quoted-scalar'
+    case '|':
+    case '>':
+      return 'block-scalar-header'
+  }
+  return null
+}
